docs: Update description
The year is 2023 and saying that something is written in Go isn't as
cool as it once was. Instead of emphasizing the language an application
was written in, it's better to assume that all modern applications are
already written in Rust, Go, and similar languages.

This change hopefully removes the ambiguity around "cover thumbnailer"
and helps emphasize that go-thumbnailer changes the desktop experience.